requirements.txt
aiohttp==3.8.3
boto3>=1.26.45
botocore>=1.29.45
python-dotenv>=0.21.0
requests==2.28.1
opensearch-py==2.0.1
Markdown~=3.4.1
